add lazy-load and module for components

+++ front/src/app/drop-field/drop-field.component.ts
@@ -3,10 +3,10 @@ import { CdkDragDrop } from '@angular/cdk/drag-drop';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { Element } from '../Element';
-import { ShareServiceService } from '../share-service.service';
-import { TransferService } from '../transfer.service';
-import { ConfirmModalComponent } from '../confirm-modal/confirm-modal.component';
+import { Element } from '../shared/interfaces/element';
+import { ShareService } from '../shared/services/shared-service/share.service';
+import { TransferService } from '../shared/services/transfer-sevice/transfer.service';
+import { ConfirmModalComponent } from './confirm-modal/confirm-modal.component';
 
 @Component({
   selector: 'app-drop-field',
@@ -15,16 +15,21 @@ import { ConfirmModalComponent } from '../confirm-modal/confirm-modal.component'
 })
 export class DropFieldComponent implements OnInit, OnDestroy {
   public dropElements: Element[] = [];
+
   public currentElement?: any = {};
+
   public objStyle?: any = {};
+
   private unsubscribeAll: Subject<any> = new Subject<any>();
-  public deleteId?: any;
+
+  public deleteId?: number;
+
   public dialogRef?: MatDialogRef<ConfirmModalComponent>;
 
   constructor(
-    private share: ShareServiceService,
+    private share: ShareService,
     private transfer: TransferService,
-    public dialog: MatDialog
+    public dialog: MatDialog,
   ) {}
 
   ngOnInit(): void {
@@ -38,20 +43,20 @@ export class DropFieldComponent implements OnInit, OnDestroy {
     this.deleteId = element.id;
   }
 
-  deleteElement(id: number) {
+  deleteElement(id: number): void {
     this.dropElements = this.dropElements.filter(
-      (item: Element) => item.id !== id
+      (item: Element) => item.id !== id,
     );
   }
 
-  onDrop(event: CdkDragDrop<Element[]>) {
+  onDrop(event: CdkDragDrop<Element[]>): void {
     this.share.drop(event);
   }
 
-  openDialog() {
+  openDialog(): void {
     this.dialogRef = this.dialog.open(ConfirmModalComponent);
     this.dialogRef.afterClosed().subscribe((result) => {
-      if (result) {
+      if (result && this.deleteId) {
         this.deleteElement(this.deleteId);
         this.transfer.deleteElementId(this.deleteId);
       }
